Accept Bitbucket user info in builder
This allows the builder to access private repositories on Bitbucket.

IMPORTANT: This means the resulting intermediary builder images used for
extraction should *NEVER* be shared (and thus never pushed onto a
registry) if such credentials were provided, unless the environment
variable is somehow cleared *AND* the layers are properly squashed.