Disable all pro plugins by default
We enable pro plugins explicitly only if their license string is
available.

MSSQL and AZSTOR didn't give us trouble since they would be loaded
without configuration and would default to not run unless their "Enable"
setting was set.  They were still loaded unnecessarily although
harmlessly.

WVP however reuses the same configuration schema as the WVB.  Even
though the latter was disabled (i.e. moved to -disabled) during the
creation of the Docker image, its default configuration would remain.
This resulted in a hard crash if the license string was not provided
since the WVP would load, be enabled by the WVB config, and fail as it
would not find a valid license (bringing Orthanc down with it).

In fact, there doesn't seem to be an "Enabled" setting for the WVB/WVP,
so for that reason and in order to avoid both spurrious messages and the
unnecessary loading of unused plugins, we disable all pro plugins by
default (since they all require a license).

Only if a license is provided would the plugins be configured and
enabled (i.e. moved to the plugins directory).

If an invalid license is provided the plugins are still configured and
enabled and thus won't silently fail to load -- instead Orthanc will
stop (failing early here is desirable) and the operator should find
helpful messages in the log.