Allow specifying WVP license string
Note that we don't systematically disable the WVB so that we can simply
fall back to it if no license string is provided for the WVP.  In the
future we should allow users to explicitly enabled/disable plugins.

Only if the WVP license string is available do we (1) reconfigure the
viewer by providing the license string and disabling the Liveshare and
(2) disable the WVB and enable the WVP.

I expect we'll want the Liveshare setting in a separate file later so
that we can override it easily (using docker-config or using an overlay
as usual).