Extract getenv logic
Notice we don't unset env vars anymore. This was a bit useless, the
setup procedure executor runs in a dedicated process. This might change
actually (we might want to source it from the docker entrypoint) but
we'll want to make sure we handle it right first.

Important: There is bash behavior I can't put my fingers on right now:
if we unset $name_$setting in the getenv function it will indeed not be
defined within the activation frame of a getenv call but will remain
defined outside of it (caller context). these bindings are supposed to
be global, so I don't know what gives. In any case, don't want to fight
with that kind of behavior, better find a way to do poor-man D-I within
bash instead (global state is a mess in all environments).