Add "standard" *_BUNDLE_DEFAULTS flag
This one forces the setup procedure to generate its assigned
configuration file regardless of whether there are relevant environment
variables set or not.  This allows end-users to use the bundle defaults
instead of the Orthanc defaults.

Important: The bundle still sets special configuration regardless of
whether BUNDLE_DEFAULTS is set to true or not (e.g. allow remote access,
default storage directory). These are things that would otherwise make
it impractical to use within a Docker container running the bundle
image.  It is likely we'll provide setup procedures to disable even
these exceptionnally in the future.