Disable all plugins by default
Important: There can still be a setup procedure for them that defines
default=true to enable them by default.  However, the end-user can now
override this behavior to disable them explicitly with the
FOOPLUGIN_BUNDLE_DEFAULTS=false environment variable (to prevent
enabling it and generating its configuration) or FOOPLUGIN_ENABLED=false
(to disable it but still maybe generate the configuration of the setup
procedure).