Replace static conf with generated conf
All plugins are now disabled by default.

Enable by:
- Adding relevant configuration file ('conf' parameter) in base layer.
- Set any environment variable for the plugin.
- Set FOOPLUGIN_ENABLED=true if one wants to enable the plugin without
generating a configuration file (which means we'll use the plugin defaults).
- Set FOOPLUGIN_BUNDLE_DEFAULTS=true if one wants to enable the plugin
and generate a configuration file using the bundle defaults.

Look for "default=true" setup procedure parameters to determine what
configuration is still used by default even if the end-user doesn't do
any of the above.  The user can now also override this using environment
variables (to disable these bundle defaults entirely or just override
certain settings).  Usually setup procedures marked 'default' are marked
so for good reason: we need to listen to addresses other than localhost
to reach the server inside the container, we really want a good default
for the plugins directory (unless we want to bypass the entire setup
mechanism) and we want a good default for the fs-based storage in case
no index/storage plugin is used.