pgsql: Fix password configuration
Empty password triggers a codepath where the PG client attempts to
connect to a local PG instance via UDS instead of not sending a
password.  Fair enough, we'll really want to abstract away the JSON
serialization instead of doing it via basic templates like this.

For now we do the ugly thing and just use an empty string for the entire
property line if we don't specify a password.