Fix double substitution
d7454d85ad0517d29f4797258a304fb8fd065cf5 didn't introduce a delay as was
clearly intended, but /another/ expansion.

Note that we also delay expansion of the filename of secrets.  This is
mostly for consistency (why do it somewhere but not everywhere?), but
also because not doing so would probably be difficult to do right as
well (since the quotes wouldn't be parsed as part of the subshell
evaluation but on the first pass instead).  Interesting point: we
anticipated that one but shellcheck catches it; this analyzer is
definitely very impressive.

For completeness: Not sure how but it *is* probably possible to expand
everything in a single pass only (leaving only a simple instruction with
only literals to eval) but again this seems very difficult to do
(ensuring no special characters will be evaluated wrong in the values,
etc).  This way feels vastly simpler (if a bit ugly to read because of
the escaping).