Build pro image on latest base image
This essentially decouples the builds; we can iterate on each separately
and in parallel without having to rebuild the other whenever they change
simply by using -n, -o/-p and omitting -l.  Of course, official releases
must take care to tag the base image as latest (with -l) to make sure
they are synchronized (this is existing practice already).