Always link orthanc-pro to "current" orthanc base
This reverts 58583bba5e11bf7c0e62d5c9c225791e6c064136 per
https://bitbucket.org/osimis/orthanc-builder/pull-requests/23/improved-build-procedure/diff#comment-47014783.

It is still possible to build only the -pro image independently though
it does not seem possible to iterate on both the base image and the -pro
image in parallel without the base image affecting the -pro image.  This
might have been an edge case we will not encounter too much again.

In any case, the use of either latest or current here will probably have
to be revisited for CI later.