azstor, mssql, wvp: Fail early on missing settings
Admittedly the plugins themselves should have proper log output if
they're getting empty strings here, but we'd rather fail early instead
of passing these empty strings.

It would also be correct to omit the settings from the configuration
file if they're not defined (and let the plugin fail appropriately) but
the way we generate the files makes it a little annoying (see password
field for PG for example).

Note that this also allows Orthanc to start without the plugin, or even
to start *with* the plugin (but Orthanc defaults) if explicitly enabled
via *_ENABLED.