res-stability: Rename setup procedure
Replace dash with underscore.

This is important because dashes aren't valid variable identifiers in
bash, nor valid environment variable identifiers in POSIX.

setup.sh:getenv would thus fail (unverified, but should).

Since the only setting in that setup procedure was a global setting,
this patch actually has no impact: the setup procedure name was never
used as a variable prefix.  Still, just in case we add normal settings
later (and to make it clear we heavily discourage such characters in
setup procedure names), we go forward with this change.