listener: Deprecate LISTEN_ALL_ADDR
The RemoteAccessAllowed Orthanc setting doesn't do the obvious thing: it
doesn't select which network interface addresses to bind to. Instead, it
filters *each* request based on the peer address. Specifically, it looks
at whether it is 127.0.0.1 or not. This explains why other loopback
addresses won't work when it is disabled, and possibly explains why
sometimes localhost resolution wouldn't work (depending on the resolver
and OS).

orthanc:Core/HttpServer/MongooseServer.cpp

  #if ORTHANC_ENABLE_MONGOOSE == 1
      static const long LOCALHOST = (127ll << 24) + 1ll;
      localhost = (request->remote_ip == LOCALHOST);
  #elif ORTHANC_ENABLE_CIVETWEB == 1
      // The "remote_ip" field of "struct mg_request_info" is tagged as
      // deprecated in Civetweb, using "remote_addr" instead.
      localhost = (std::string(request->remote_addr) == "127.0.0.1");
  #else
  #error
  #endif

      // Check remote calls
      if (!server.IsRemoteAccessAllowed() &&
          !localhost)
      {
        output.SendUnauthorized(ORTHANC_REALM);
        return;
      }

This means the name of the setting is confusing. We'll be introducing
another setup procedure with a more obvious name in a subsequent commit.

Importantly, we determined that it is in fact impossible to select the
network interface address to bind to in Orthanc Server. The Mongoose
Server abstraction in Orthanc merely accepts a port number, and it is
never combined with an address to form a Mongoose "port specification".
In Orthanc's defense, Mongoose calling it a "port specification" doesn't
make it obvious that an optional address can be prefixed to it.

When no network interface address is specified to the Mongoose port
specification, 0.0.0.0 will be used (the socket struct containing a
union of the sockaddr_* structs is zeroed-out), which ultimately means
that Orthanc will *always* bind to all network interface addresses.
Thus, we cannot replace this functionality and are forced to deprecate
it until Orthanc is patched to support it. If somebody relied on it, it
never worked the way they thought it would.

Notice: The "deprecated" setup procedure parameter currently has no
effect, but the setup procedure executor might log a message about it in
the future.