Add gcp-dicom setup procedure
While we do not have facilities to handle dynamic environment variables
(e.g. to create complex lists), it is still useful to reify as many
settings as possible as environment variables so that higher-level
tooling can handle them in isolation (among other things).  So, we
decide to establish a "single account" mode which should cover the vast
majority of cases, and still offer an escape hatch in the form of the
ACCOUNTS setting that allows specifying everything in a single JSON
object (pending a better solution, hopefully not implemented with the
setup procedures model).

We've also clearly reached the limits of the plaintext-based templating
approach; this is about the best I could do without a JSON serializer
(and a language capable of properly expressing rich data structures).

Notice: Adding a "-DICOM" suffix to the name.  It's quite likely we'll
have other plugins related to the GCP; for example a plugin to leverage
GCP Object Storage as an Orthanc Storage backend.  We anticipate
confusion around a plugin simply named "GCP" in that potential
situation, and a possible rename of the plugin.  We hope to get ahead of
this here.

Resolves OPS-442