Link or copy plugins to best location
To have a read-only topmost container layer we want to mount volumes or
temporary filesystems to all writeable locations of the tree.  In order
to avoid an unnecessary copy from the relevant image layer into the new
filesystems during startup we attempt to mount them only on empty
directories.

This patch prevents the write to delete the link from the
plugins-disabled directory.

We do this by copying the shared objects instead of
renaming/relinking/moving them.  This is necessary given we want to
support mounting these paths on different filesystems, as explained
above (bindmounts, tmpfs, etc).

For the base case (which remains for backward-compatibility with
containers that still need a writable topmost layer) we remark that
because of the copy-on-write semantics of writing to the topmost layer,
the unlinking of the files is actually superfluous.  We thus simply link
them (a move/rename is essentially a link-to-new-dir and
unlink-from-old-dir, we just skip the latter).